YT_Transcript/requirements.txt
youtube_transcript_api
google_generativeai
python_dotenv
pathlib